pyproject.toml
[project]
name = "logistic-audit-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Gaurav",email = "gauravgowda452@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0.0"
dependencies = [
    "langgraph (>=1.1.8,<2.0.0)",
    "fastapi (>=0.136.0,<0.137.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pydantic (>=2.13.3,<3.0.0)",
    "pydantic-settings (>=2.14.0,<3.0.0)",
    "neo4j (>=6.1.0,<7.0.0)",
    "langchain-core (>=1.3.0,<2.0.0)",
    "anthropic (>=0.96.0,<0.97.0)",
    "openai (>=2.32.0,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "google-genai (>=1.73.1,<2.0.0)"
]

[tool.poetry]
packages = [
    { include = "app" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
